Gave up on returning a special String.
Turns out Javascript has 2 strings, one primitive, one that is an
object. I was using the object one in order to add properties to it
(one can't add properties to the primitive string), but gave up on
this because jquery tests whether typeof(something) == 'string',
and jquery.text(someString) was not working anymore.